requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
DateTime==4.4
fastapi==0.78.0
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
mysql-connector-python==8.0.29
numpy==1.22.4
packaging==21.3
pandas==1.4.3
pluggy==1.0.0
protobuf==4.21.1
py==1.11.0
pycodestyle==2.8.0
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
zope.interface==5.4.0
